refactor(puppeteer): remove dead code identified by knip
Delete unused files:
- utils/uploadMetrics.js (at wrong path, unused)
- src/shared/constants/ (4 files, no imports)
- src/shared/utils/textFormatter.js (no imports)
- src/domains/automation/utils/humanBehaviorManager.js (no imports)

Fix broken import paths:
- linkedinService.js: crypto.js path fixed
- linkedinInteractionController.js: interactionQueue.js path fixed
- linkedinContactService.js: textExtractionService.js path fixed

Remove unused dependencies:
- @aws-sdk/client-dynamodb
- @aws-sdk/lib-dynamodb

Refactor s3TextUploadService.js to use inline metrics
