+++ puppeteer/package.json
@@ -14,9 +14,7 @@
     "test:coverage": "vitest run --coverage"
   },
   "dependencies": {
-    "@aws-sdk/client-dynamodb": "^3.844.0",
     "@aws-sdk/client-s3": "^3.832.0",
-    "@aws-sdk/lib-dynamodb": "^3.844.0",
     "@google/generative-ai": "^0.24.0",
     "axios": "^1.11.0",
     "cors": "^2.8.5",
