--- puppeteer/src/shared/constants/http.js
@@ -1,22 +1,0 @@
-// HTTP status codes
-export const HTTP_STATUS = {
-  OK: 200,
-  CREATED: 201,
-  ACCEPTED: 202,
-  NO_CONTENT: 204,
-  BAD_REQUEST: 400,
-  UNAUTHORIZED: 401,
-  FORBIDDEN: 403,
-  NOT_FOUND: 404,
-  CONFLICT: 409,
-  UNPROCESSABLE_ENTITY: 422,
-  INTERNAL_SERVER_ERROR: 500,
-  BAD_GATEWAY: 502,
-  SERVICE_UNAVAILABLE: 503,
-};
-
-export const CONTENT_TYPES = {
-  JSON: 'application/json',
-  TEXT: 'text/plain',
-  HTML: 'text/html',
-};
