--- puppeteer/src/shared/constants/index.js
@@ -1,4 +1,0 @@
-// Backend constants barrel export
-export * from './http.js';
-export * from './linkedin.js';
-export * from './limits.js';
