--- puppeteer/src/shared/constants/linkedin.js
@@ -1,30 +1,0 @@
-// LinkedIn-specific constants
-export const LINKEDIN_SELECTORS = {
-  // Authentication
-  EMAIL_INPUT: '#username',
-  PASSWORD_INPUT: '#password',
-  LOGIN_BUTTON: '.btn__primary--large',
-
-  // Connections
-  CONNECTIONS_LIST: '.mn-connection-card',
-  MESSAGE_BUTTON: '.message-anywhere-button',
-  CONNECT_BUTTON: '.artdeco-button--secondary',
-
-  // Profile
-  PROFILE_NAME: '.text-heading-xlarge',
-  PROFILE_HEADLINE: '.text-body-medium',
-  ABOUT_SECTION: '#about',
-
-  // Messages
-  MESSAGE_THREAD: '.msg-conversation-listitem',
-  MESSAGE_INPUT: '.msg-form__contenteditable',
-  SEND_BUTTON: '.msg-form__send-button',
-};
-
-export const LINKEDIN_URLS = {
-  BASE: 'https://www.linkedin.com',
-  LOGIN: 'https://www.linkedin.com/login',
-  FEED: 'https://www.linkedin.com/feed',
-  CONNECTIONS: 'https://www.linkedin.com/mynetwork/invite-connect/connections',
-  MESSAGES: 'https://www.linkedin.com/messaging',
-};
